# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UniGate API",
    "version": "0.1.0",
    "description": "Vehicle hiring & management platform. Money is a decimal string; timestamps are ISO-8601 UTC; every response uses the same envelope."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "code",
              "requestId"
            ]
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "error"
        ]
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "documentTypeCode": {
            "type": "string"
          },
          "target": {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "kind",
              "id"
            ]
          },
          "originalFilename": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer"
          },
          "checksumSha256": {
            "type": "string"
          },
          "uploadStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "UPLOADED",
              "FAILED",
              "QUARANTINED"
            ]
          },
          "verificationStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED",
              "EXPIRED"
            ]
          },
          "verifiedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "rejectionReason": {
            "type": [
              "string",
              "null"
            ]
          },
          "issueDate": {
            "type": [
              "string",
              "null"
            ]
          },
          "expiryDate": {
            "type": [
              "string",
              "null"
            ]
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "INTERNAL",
              "SHARED_WITH_COUNTERPARTY"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "documentTypeCode",
          "target",
          "originalFilename",
          "mimeType",
          "sizeBytes",
          "checksumSha256",
          "uploadStatus",
          "verificationStatus",
          "verifiedAt",
          "rejectionReason",
          "issueDate",
          "expiryDate",
          "visibility",
          "createdAt",
          "updatedAt"
        ],
        "description": "Metadata only — never the storage key or a URL."
      },
      "DocumentEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/Document"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "meta": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "meta"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/documents/{id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Metadata",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Soft delete (documents.delete); refused when an approval decision references it",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}